sandbox: Move the static/wall-clock CLI tests downstream.
They're no longer "common" options.
