+++ ledger/sandbox/src/test/suite/scala/platform/sandboxnext/CliSpec.scala
@@ -8,6 +8,7 @@ import com.daml.ledger.participant.state.v1.SeedService.Seeding
 import com.daml.lf.data.Ref
 import com.daml.platform.common.LedgerIdMode
 import com.daml.platform.sandbox.cli.CommonCliSpecBase
+import com.daml.platform.services.time.TimeProviderType
 
 class CliSpec extends CommonCliSpecBase(Cli) {
 
@@ -23,6 +24,23 @@ class CliSpec extends CommonCliSpecBase(Cli) {
     "parse the sql-backend-jdbcurl flag when given" in {
       val jdbcUrl = "jdbc:postgresql://localhost:5432/test?user=test"
       checkOption(Array("--sql-backend-jdbcurl", jdbcUrl), _.copy(jdbcUrl = Some(jdbcUrl)))
+    }
+
+    "apply static time when given" in {
+      checkOption(Array("-s"), _.copy(timeProviderType = Some(TimeProviderType.Static)))
+      checkOption(Array("--static-time"), _.copy(timeProviderType = Some(TimeProviderType.Static)))
+    }
+
+    "apply wall-clock time when given" in {
+      checkOption(
+        Array("--wall-clock-time"),
+        _.copy(timeProviderType = Some(TimeProviderType.WallClock)))
+      checkOption(Array("-w"), _.copy(timeProviderType = Some(TimeProviderType.WallClock)))
+    }
+
+    "return None when both static and wall-clock time are given" in {
+      val config = cli.parse(requiredArgs ++ Array("--static-time", "--wall-clock-time"))
+      config shouldEqual None
     }
 
     "parse the contract-id-seeding mode when given" in {
